api/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
base58==2.1.1
bitarray==2.6.0
bitcoin-python==0.3
bitcoinlib==0.6.12
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.0
coincurve==18.0.0
constantly==15.1.0
cryptography==41.0.4
cryptomus==1.1
cytoolz==0.12.2
daphne==4.0.0
Django==4.2.5
django-jazzmin==2.6.0
djangochannelsrestframework==1.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
eth-abi==4.2.1
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.0
eth-utils==2.2.1
exceptiongroup==1.1.3
fastecdsa==2.3.0
frozenlist==1.4.0
graphql-core==3.2.3
graphql_query==1.0.3
greenlet==3.0.0
h11==0.14.0
hexbytes==0.3.1
httpcore==0.18.0
httpx==0.25.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
ipaddress==1.0.23
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
MarkupSafe==2.1.3
mnemonic==0.20
monero==1.1.1
multidict==6.0.4
numpy==1.26.0
orjson==3.9.9
packaging==23.2
parsimonious==0.9.0
pluggy==1.3.0
protobuf==4.24.4
psycopg2-binary==2.9.9
py-solc==3.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
pydantic==1.10.13
PyNaCl==1.5.0
pyOpenSSL==23.2.0
PySocks==1.7.1
pytest==7.4.2
pytest-asyncio==0.21.1
python-bitcoinlib==0.12.2
python-bitcoinrpc==1.0
python-dotenv==1.0.0
pytonlib==0.0.50
pytz==2023.3.post1
pyunormalize==15.0.0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rlp==3.0.0
rpds-py==0.10.6
semantic-version==2.10.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
tomli==2.0.1
ton==0.26
ton-client-py==1.44.3.0
tonsdk==1.0.13
TonTools==2.1.2
toolz==0.12.0
tronpy==0.4.0
tvm-valuetypes==0.0.10
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6
varint==1.0.2
web3==6.11.0
websockets==11.0.3
yarl==1.9.2
zope.interface==6.0
